requirements.txt
beautifulsoup4==4.5.3
conda==4.3.16
google-auth==1.20.1
google-auth-oauthlib==0.4.1
greenlet==0.4.11
gspread==3.6.0
numpy==1.16.3
oauth2client==4.1.3
pandas==0.24.2
python-dotenv==0.14.0
requests==2.12.4
urllib3==1.25.9
